Able to pass items as link state to other component
Appears that must wrap both sender and caller in withRouter to be
able to pass state with router link object if using Redux connected
compoenent. Otherwise you can see the state in the route object in
React dev tools, but you can't access it within the component. Need
to experiment more with this and map out the combinations to see how
it all works.